Clean up and Refactoring for obsolete files
