--- src/test/java/com/sparta/team/driver/ChromeDriverManager.java
@@ -1,43 +1,0 @@
-package com.sparta.team.driver;
-
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeDriverService;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.remote.DesiredCapabilities;
-
-import java.io.File;
-
-public class ChromeDriverManager extends DriverManager {
-
-    private ChromeDriverService chService;
-
-    @Override
-    public void startService() {
-        if (null == chService) {
-            try {
-                chService = new ChromeDriverService.Builder()
-                        .usingDriverExecutable(new File("chromedriver.exe"))
-                        .usingAnyFreePort()
-                        .build();
-                chService.start();
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
-    @Override
-    public void stopService() {
-        if (null != chService && chService.isRunning())
-            chService.stop();
-    }
-
-    @Override
-    public void createDriver() {
-        DesiredCapabilities capabilities = DesiredCapabilities.chrome();
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("test-type");
-        capabilities.setCapability(ChromeOptions.CAPABILITY, options);
-        driver = new ChromeDriver(chService, capabilities);
-    }
-}
