--- src/test/java/com/sparta/team/driver/DriverManager.java
@@ -1,26 +1,0 @@
-package com.sparta.team.driver;
-
-import org.openqa.selenium.WebDriver;
-
-public abstract class DriverManager {
-
-    protected WebDriver driver;
-    protected abstract void startService();
-    protected abstract void stopService();
-    protected abstract void createDriver();
-
-    public void quitDriver() {
-        if (null != driver) {
-            driver.quit();
-            driver = null;
-        }
-    }
-
-    public WebDriver getDriver() {
-        if (null == driver) {
-            startService();
-            createDriver();
-        }
-        return driver;
-    }
-}
