--- src/test/java/com/sparta/team/driverManager/DriverType.java
@@ -1,7 +1,0 @@
-package com.sparta.team.driverManager;
-
-public enum DriverType {
-    CHROME,
-    FIREFOX,
-}
-
